refactor: split pack 7
