+++ websites/J/Jackbox/presence.ts
@@ -52,29 +52,6 @@ let gamePlayerState: GamePlayerState = {
 	browsingTimestamp = Math.round(Date.now() / 1000),
 	gametag: string;
 
-const uploadedFiles: Record<string, string> = {};
-async function uploadFile(url: string, defaultImage: string): Promise<string> {
-	if (uploadedFiles[url]) return uploadedFiles[url];
-	uploadedFiles[url] = defaultImage;
-
-	try {
-		const data = await fetch(url).then(res => res.blob()),
-			resultURL = await fetch("https://bashupload.com/", {
-				method: "POST",
-				body: data,
-			})
-				.then(res => res.text())
-				.then(text => text.match(/https(.*)/)?.[0]);
-
-		presence.info(resultURL);
-		uploadedFiles[url] = resultURL;
-		return resultURL;
-	} catch (err) {
-		presence.error(err);
-		return url;
-	}
-}
-
 if (window.location.hostname === "jackbox.tv") {
 	setInterval(async () => {
 		const playerStateLogs = await presence.getLogs(
@@ -143,27 +120,6 @@ if (window.location.hostname === "jackbox.tv") {
 }
 
 const Games: Record<string, Game> = {
-	// Party Pack 7
-	quiplash3: {
-		name: "Quiplash 3",
-		logo: "https://i.imgur.com/Jds4R56.png",
-	},
-	"jackbox-talks": {
-		name: "Talking Points",
-		logo: "https://i.imgur.com/8Amy9Di.png",
-	},
-	"blanky-blank": {
-		name: "Blather 'Round",
-		logo: "https://i.imgur.com/qLW2Fft.png",
-	},
-	everyday: {
-		name: "The Devils and the Details",
-		logo: "https://i.imgur.com/dcJq65O.png",
-	},
-	worldchamps: {
-		name: "Champ'd Up",
-		logo: "https://i.imgur.com/KnAS7yD.png",
-	},
 	// Party Pack 8
 	"the-wheel": {
 		name: "The Wheel of Enormous Proportions",
@@ -242,351 +198,6 @@ presence.on("UpdateData", async () => {
 				}
 				if (useDetails) {
 					switch (game) {
-						// Party Pack 7
-						case Games.quiplash3: {
-							presenceData.smallImageKey = getComputedStyle(
-								document.querySelector<HTMLDivElement>("#playericon")
-							).backgroundImage.match(/^url\("(.*)"\)$/)?.[1];
-							switch (gamePlayerState.state) {
-								case "Lobby": {
-									presenceData.state = "Waiting in lobby";
-									break;
-								}
-								case "Logo": {
-									presenceData.state = "Waiting";
-									break;
-								}
-								case "EnterSingleText": {
-									presenceData.state = "Answering a prompt";
-									break;
-								}
-								case "MakeSingleChoice": {
-									presenceData.state = "Voting for their favorite answer";
-									break;
-								}
-								case "EnterTextList": {
-									presenceData.state = "Answering a Thriplash prompt";
-									break;
-								}
-								default: {
-									if (gamePlayerState.validActions) {
-										switch (
-											(gamePlayerState.validActions as string[]).join(",")
-										) {
-											case "toggle-visibility,new,load,exit": {
-												presenceData.state = "In the Custom Content menu";
-												break;
-											}
-											case "title,close": {
-												presenceData.state = "Naming a custom Quiplash episode";
-												break;
-											}
-											case "add,toggle-visibility,close":
-											case "add,remove,toggle-visibility,done": {
-												presenceData.state =
-													"Adding prompts to a custom Quiplash episode";
-												break;
-											}
-											case "submit,unlock,toggle-visibility,play,remove-content,episodes": {
-												presenceData.state =
-													"Viewing a custom Quiplash episode";
-												break;
-											}
-										}
-									}
-								}
-							}
-							break;
-						}
-						case Games["jackbox-talks"]: {
-							switch (gamePlayerState.state) {
-								case "Lobby": {
-									presenceData.state = "Waiting in lobby";
-									break;
-								}
-								case "Logo": {
-									presenceData.state = "Waiting";
-									break;
-								}
-								case "Camera": {
-									presenceData.state = "Taking a profile picture";
-									break;
-								}
-								case "EnterSingleText": {
-									const { entryId } = gamePlayerState;
-									if (entryId.startsWith("prompt"))
-										presenceData.state = "Creating topics";
-									else if (entryId === "WriteQuote")
-										presenceData.state = "Writing a quote about the talk";
-									else if (entryId === "NameAward")
-										presenceData.state = "Naming an award";
-									break;
-								}
-								case "Awards": {
-									presenceData.state = "Giving out their award";
-									break;
-								}
-								case "MakeSingleChoice": {
-									const { classes, choices, prompt } = gamePlayerState;
-									if (
-										prompt.html === "PICK THE TITLE OF THE TALK YOU WILL GIVE"
-									)
-										presenceData.state = "Choosing a talk title";
-									else {
-										switch (classes[0]) {
-											case "SkipTutorial": {
-												presenceData.state = "Watching the tutorial";
-												break;
-											}
-											case "Presenter": {
-												if (prompt.html.startsWith("RATE HOW WELL "))
-													presenceData.state = "Rating their assistant";
-												else if (prompt.text === "THANK YOU.") {
-													presenceData.state =
-														"Presenting their talk - thank you";
-												} else {
-													presenceData.state =
-														"Presenting their talk - preparation";
-												}
-												break;
-											}
-											case "Assistant": {
-												if (
-													prompt.html ===
-													"PICK THE BEST PICTURE TO REPRESENT THE TALK"
-												) {
-													presenceData.state =
-														"Choosing a picture for the talk";
-												} else presenceData.state = "Assisting their presenter";
-												break;
-											}
-											default: {
-												if (
-													(choices as { className: string }[])[0].className ===
-													"voteUp"
-												)
-													presenceData.state = "Reacting to the speech";
-											}
-										}
-									}
-									break;
-								}
-								case "Draw": {
-									presenceData.state = "Presenting their talk - slide";
-									break;
-								}
-							}
-							break;
-						}
-						case Games["blanky-blank"]: {
-							switch (gamePlayerState.state) {
-								case "Lobby": {
-									presenceData.state = "Waiting in lobby";
-									break;
-								}
-								case "Logo": {
-									presenceData.state = "Waiting";
-									break;
-								}
-								case "MakeSingleChoice": {
-									switch (gamePlayerState.choiceType) {
-										case "skipTutorial": {
-											presenceData.state = "Watching the tutorial";
-											break;
-										}
-										case "password": {
-											presenceData.state = "Choosing a prompt";
-											break;
-										}
-									}
-									break;
-								}
-								case "MakeSentence": {
-									switch ((gamePlayerState.sentence as { type: string }).type) {
-										case "writing": {
-											presenceData.state = "Crafting initial sentence";
-											break;
-										}
-										case "call": {
-											presenceData.state = "Crafting a sentence";
-											break;
-										}
-										case "response": {
-											presenceData.state =
-												"Crafting a sentence using players' guesses";
-											break;
-										}
-										case "mybad": {
-											presenceData.state =
-												"Deciding if they should have known the answer";
-											break;
-										}
-									}
-									break;
-								}
-								case "EnterSingleText": {
-									presenceData.state = "Guessing the object";
-									break;
-								}
-							}
-							break;
-						}
-						case Games.everyday: {
-							switch (gamePlayerState.state) {
-								case "Lobby": {
-									presenceData.state = "Waiting in lobby";
-									break;
-								}
-								case "Logo": {
-									presenceData.state = "Waiting";
-									break;
-								}
-								case "MakeSingleChoice": {
-									switch (gamePlayerState.choiceId) {
-										case "ChangeCharacter": {
-											presenceData.state = "Choosing a character";
-											break;
-										}
-										default: {
-											if (gamePlayerState.choiceType === "SkipIntro")
-												presenceData.state = "Watching the intro";
-											else if (gamePlayerState.choiceType === "ShowTutorial") {
-												presenceData.state =
-													"Deciding if they want to watch the tutorial";
-											}
-										}
-									}
-									break;
-								}
-								case "CancelerMechanic": {
-									presenceData.state =
-										"Blocking a member from their selfish task";
-									break;
-								}
-								case "TaskList": {
-									presenceData.state = "Choosing a task";
-									break;
-								}
-								case "NothingMechanic": {
-									presenceData.state = "Doing nothing";
-									break;
-								}
-								case "ScrubMechanic": {
-									presenceData.state = "Scrubbing something";
-									break;
-								}
-								case "RotateMechanic": {
-									presenceData.state = "Rotating something";
-									break;
-								}
-								case "TapMechanic": {
-									presenceData.state = "Tapping something";
-									break;
-								}
-								case "SwipeMechanic": {
-									presenceData.state = "Swiping something";
-									break;
-								}
-								case "TravelMechanicParticipant": {
-									presenceData.state = "Travelling as a passenger";
-									break;
-								}
-								case "TravelMechanicCaptain": {
-									presenceData.state = "Travelling as the driver";
-									break;
-								}
-								case "PhoneMechanic": {
-									presenceData.state = "Calling someone";
-									break;
-								}
-								case "DialogMechanic": {
-									presenceData.state = "Talking to someone";
-									break;
-								}
-								case "SearchMechanicCaptain":
-								case "SearchMechanicParticipant": {
-									presenceData.state = "Searching for something";
-									break;
-								}
-								case "TextMechanicParticipant": {
-									presenceData.state = "Writing a text";
-									break;
-								}
-								case "TextMechanicCaptain": {
-									presenceData.state = "Reading a text";
-									break;
-								}
-								case "InstructionsMechanicCaptain": {
-									presenceData.state = "Following instructions";
-									break;
-								}
-								case "InstructionsMechanicParticipant": {
-									presenceData.state = "Giving instructions";
-									break;
-								}
-								case "HoldMechanic": {
-									presenceData.state = "Holding something";
-									break;
-								}
-							}
-							break;
-						}
-						case Games.worldchamps: {
-							presenceData.smallImageKey = getComputedStyle(
-								document.querySelector("#playericon")
-							).backgroundImage.match(/^url\("(.*)"\)$/)?.[1];
-							switch (gamePlayerState.state) {
-								case "Lobby": {
-									presenceData.state = "Waiting in lobby";
-									break;
-								}
-								case "Logo": {
-									presenceData.state = "Waiting";
-									break;
-								}
-								case "MakeSingleChoice": {
-									const { choiceId } = gamePlayerState;
-									if (choiceId) {
-										if (choiceId.startsWith("FlipChoice"))
-											presenceData.state = "Flipping their character";
-									} else if (
-										gamePlayerState.prompt.html?.startsWith("Who is <br>")
-									)
-										presenceData.state = "Voting for the best champion";
-									else if (
-										gamePlayerState.prompt.html?.startsWith(
-											"Swap your character or keep it in<br>"
-										)
-									) {
-										presenceData.state =
-											"Choosing whether to swap their character";
-									} else presenceData.state = "Watching the tutorial";
-									break;
-								}
-								case "Draw": {
-									const { entryId } = gamePlayerState;
-									if (entryId.startsWith("champion")) {
-										presenceData.state = "Drawing a champion";
-										break;
-									} else if (entryId.startsWith("challenger")) {
-										const imageLink =
-											document.querySelector<HTMLImageElement>(
-												".imageData"
-											)?.src;
-										if (imageLink) {
-											presenceData.largeImageKey = await uploadFile(
-												imageLink,
-												logo
-											);
-										}
-										presenceData.state = "Drawing a challenger";
-										break;
-									}
-									break;
-								}
-							}
-							break;
-						}
 						// Party Pack 8
 						case Games["the-wheel"]: {
 							switch (gamePlayerState.kind) {
